backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
redis
qdrant-client>=1.12.0
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
PyMuPDF
python-multipart
python-jose[cryptography]
passlib[bcrypt]
groq
httpx
fakeredis
google-auth
gunicorn
python-dotenv
fastapi-mail
tavily-python
langchain-tavily
langchain-groq
langgraph
langchain-text-splitters
python-slugify
tiktoken
